Problems sending/receiving MMS, tears down MMS APN
The MMS APN incorrectly tears down when connected. This is due to an
error in the code checking if the phone is connected to a MMS APN. In
TransactionSettings the default path is used, but this causes the
selection parameter to be ignored in TelephonyProvider. The result is
that when connected to a non-MMS APN it believes it is connected to a
MMS APN. The fix is to simply remove default in the URI and construct
the WHERE clause manually.

Change-Id: I25ffc26c68a40676841e34fd56aeab08604f4920