Make the liftcycle of cursor under the Activity
The cursors of ComposeCardActivity and ConversationList
were never be closed. So the database will throw an
CursorWindowAllocationException if the cursors which not
be closed are too many to be allocated then the database
will return a NULL cursor back to onQueryComplete().
After that Mms will be crashed by the NullPointerException.

Change-Id: Ib8730b326edf9177bbcce48a42b85405acff51ba
Signed-off-by: Roger Chen <cxr514033970@gmail.com>