QuickMessage: Prevent the screen from locking while user is typing
If the 'Wake and unlock' is set and the screen is auto unlocked, a
timer is set to relock the screen after 20 seconds.  This commit
fixes the problem where, while the user was typing, the timer would
expire and the screen would lock.  We now 'poke' the wakelock to
restart the timer on user keypress.

Change-Id: I8c76886df7ad5a6ed8513d8b2b589fd64675245c